+++ stable_baselines3/common/utils.py
@@ -4,7 +4,7 @@ import platform
 import random
 from collections import deque
 from itertools import zip_longest
-from typing import Dict, Iterable, Optional, Union
+from typing import Dict, Iterable, Optional, Tuple, Union
 
 import gym
 import numpy as np
@@ -465,23 +465,26 @@ def should_collect_more_steps(
         )
 
 
-def system_info(print_info: bool = True) -> Dict[str, str]:
+def system_info(print_info: bool = True) -> Tuple[Dict[str, str], str]:
     """
     Retrieve system and python env info for the current system.
 
     :param print_info: Whether to print or not those infos
-    :return: Dictionary summing up the version for each relevant package.
+    :return: Dictionary summing up the version for each relevant package
+        and a formatted string.
     """
     env_info = {
         "OS": f"{platform.platform()} {platform.version()}",
         "Python": platform.python_version(),
-        "stable-baselines3": stable_baselines3.__version__,
+        "Stable-Baselines3": stable_baselines3.__version__,
         "PyTorch": th.__version__,
         "GPU Enabled": str(th.cuda.is_available()),
         "Numpy": np.__version__,
         "Gym": gym.__version__,
     }
+    env_info_str = ""
+    for key, value in env_info.items():
+        env_info_str += f"{key}: {value}\n"
     if print_info:
-        for key, value in env_info.items():
-            print(f"{key}: {value}")
-    return env_info
+        print(env_info_str)
+    return env_info, env_info_str